Decouple pure beehive math + EMPTY balance from runtime deps
- Move getActiveFlower, getCurrentHoneyProduced, getCurrentSpeed, and
  areBeehivesEmpty from beehiveMachine.ts into a new pure module
  features/game/lib/beehiveProduction.ts. beehiveMachine.ts re-exports
  them so existing imports keep working. Read-only consumers can now
  import the honey math without pulling xstate in transitively.

- Drop the top-level "web3-utils" import from features/game/lib/
  constants.ts. The two callsites were fromWei("0") — identically zero
  — so the calls are replaced with new Decimal(0). The dependency was
  the only thing in this file that prevented downstream tools from
  importing recovery-time / item constants without web3-utils.

All 271 test suites pass (3,532 tests). Motivated by a downstream
project (sunflower-land-overview) that consumes pure yield/time math
from this repo and needs to avoid stubbing xstate and web3-utils.
